Breakup indieauth function and middleware. Update tests
