--- app/lib/auth/index.js
@@ -1,9 +1,0 @@
-/**
- * Verify access IndieAuth tokens and determine scope.
- *
- * @module auth
- */
-module.exports = {
-  indieauth: require(process.env.PWD + '/app/lib/auth/indieauth'),
-  scope: require(process.env.PWD + '/app/lib/auth/scope')
-};
