+++ app/lib/indieauth/check-scope.js
@@ -2,45 +2,18 @@ const {IndieKitError} = require(process.env.PWD + '/app/errors');
 const logger = require(process.env.PWD + '/app/logger');
 
 /**
- * Middleware function for checkScope
- *
- * @memberof auth
- * @module scope.middleware
- * @param {String} requiredScope Scope to check
- * @param {Object} req Express request
- * @param {Object} res Express response
- * @param {Function} next Express callback
- * @return {Function} next Express callback
- * @return {Object} Error
- */
-const middleware = requiredScope => (req, res, next) => {
-  try {
-    const tokenScope = res.locals.indieauthToken.scope;
-    module.exports.checkScope(requiredScope, tokenScope);
-    return next();
-  } catch (error) {
-    const {message} = error;
-    return res.status(message.status).json({
-      error: message.error,
-      error_description: message.error_description,
-      scope: requiredScope
-    });
-  }
-};
-
-/**
  * Checks if scope(s) in authenticated token contains required scope.
  * Automatically handles `post` and `create` as the same thing
  *
- * @memberof auth
- * @module scope.checkScope
+ * @memberof indieauth
+ * @module checkScope
  * @param {String} requiredScope Required scope
  * @param {Object} tokenScope Scope(s) provided by access token
  * @return {Boolean} True if tokenScope includes requiredScope
  */
-const checkScope = (requiredScope, tokenScope) => {
-  logger.debug('auth.scope.checkScope, required scope: %s', requiredScope);
-  logger.debug('auth.scope.checkScope, token scope: %s', tokenScope);
+module.exports = (requiredScope, tokenScope) => {
+  logger.debug('auth.checkScope, required scope: %s', requiredScope);
+  logger.debug('auth.checkScope, token scope: %s', tokenScope);
 
   if (!requiredScope) {
     throw new IndieKitError({
@@ -79,9 +52,4 @@ const checkScope = (requiredScope, tokenScope) => {
     error: 'insufficient_scope',
     error_description: `Access token does not meet requirements for requested scope (${requiredScope})`
   });
-};
-
-module.exports = {
-  middleware,
-  checkScope
 };
