+++ app/lib/micropub/create-media.js
@@ -1,4 +1,4 @@
-const auth = require(process.env.PWD + '/app/lib/auth');
+const indieauth = require(process.env.PWD + '/app/lib/indieauth');
 const logger = require(process.env.PWD + '/app/logger');
 const media = require(process.env.PWD + '/app/lib/media');
 
@@ -14,7 +14,7 @@ const media = require(process.env.PWD + '/app/lib/media');
  */
 module.exports = [
   (req, res, next) => {
-    return auth.scope.middleware('create')(req, res, next);
+    return indieauth.checkScope('create')(req, res, next);
   },
   async (req, res, next) => {
     const {file} = req;
